Agregar mapeo de imagen, cambiar nacionalidad a Short
Exponer la URL de la imagen de persona y normalizar el tipo de nacionalidad a Short en el controlador, repositorio, interfaz de servicio e implementación. Agregar @Transactional(readOnly = true) a searchJugador. Actualizar JugadorResponseDto y JugadorMapper para incluir la imagen, y ajustar la consulta/parámetros del repositorio y los tipos de parámetros del request en el controlador. Ampliar las pruebas unitarias (JugadorServiceImplTest) para cubrir escenarios de búsqueda, actualización, carga de archivos, estadísticas de carta y liberarJugador; mejorar los test data builders (JugadorTestDataBuilder, PersonaTestDataBuilder) para soportar construcción basada en requests y valores predeterminados de cartas.

+++ liga-mx/src/main/java/com/fdf/liga_mx/services/IJugadorService.java
@@ -11,7 +11,7 @@ import java.util.Map;
 import java.util.UUID;
 
 public interface IJugadorService extends CrudService<JugadorRequest, JugadorResponseDto, Jugador, Long>{
-    Page<JugadorResponseDto> searchJugador(Integer page, Integer size, String sorts, String nombre, Integer nacionalidad, Short club);
+    Page<JugadorResponseDto> searchJugador(Integer page, Integer size, String sorts, String nombre, Short nacionalidad, Short club);
     Map<String,Integer> obtenerTarjetasJugadorPorTorneoId(Long jugadorId, Long torneoId);
 
     void updateTarjetasByPartidoId(UUID id);
